Moved position utils.
