+++ src/engine/core/managers/simulation/simulation_activities.ts
@@ -16,7 +16,7 @@ import { Squad } from "@/engine/core/objects/server/squad/Squad";
 import { isInTimeInterval } from "@/engine/core/utils/game/game_time";
 import { hasAlifeInfo } from "@/engine/core/utils/info_portion";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getServerDistanceBetween } from "@/engine/core/utils/object/object_location";
+import { getServerDistanceBetween } from "@/engine/core/utils/position";
 import { isAnySquadMemberEnemyToActor } from "@/engine/core/utils/relation";
 import { communities, TCommunity } from "@/engine/lib/constants/communities";
 import { infoPortions } from "@/engine/lib/constants/info_portions";
