+++ src/engine/core/managers/simulation/utils/simulation_priority.ts
@@ -1,6 +1,6 @@
 import { TSimulationObject } from "@/engine/core/managers/simulation";
 import { Squad } from "@/engine/core/objects/server/squad";
-import { areObjectsOnSameLevel, getServerDistanceBetween } from "@/engine/core/utils/object";
+import { areObjectsOnSameLevel, getServerDistanceBetween } from "@/engine/core/utils/position";
 import { Optional, ServerObject, TDistance, TRate } from "@/engine/lib/types";
 
 /**
