+++ src/engine/core/objects/ai/state/state/ActionStateToIdle.ts
@@ -3,7 +3,7 @@ import { action_base, LuabindClass } from "xray16";
 import { StalkerStateManager } from "@/engine/core/objects/ai/state/StalkerStateManager";
 import { EStalkerState } from "@/engine/core/objects/animation/types";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { sendToNearestAccessibleVertex } from "@/engine/core/utils/object/object_location";
+import { sendToNearestAccessibleVertex } from "@/engine/core/utils/position";
 import { EClientObjectPath, TName } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
