+++ src/engine/core/objects/server/squad/Squad.ts
@@ -6,19 +6,17 @@ import {
   openLoadMarker,
   openSaveMarker,
   registerObjectStoryLinks,
+  registerSimulationObject,
   registry,
   resetStalkerState,
   SMART_TERRAIN_MASKS_LTX,
   softResetOfflineObject,
   SQUAD_BEHAVIOURS_LTX,
   SYSTEM_INI,
-} from "@/engine/core/database";
-import {
-  registerSimulationObject,
   unregisterSimulationObject,
+  unregisterStoryLinkByObjectId,
   updateSimulationObjectAvailability,
-} from "@/engine/core/database/simulation";
-import { unregisterStoryLinkByObjectId } from "@/engine/core/database/story_objects";
+} from "@/engine/core/database";
 import { EGameEvent, EventsManager } from "@/engine/core/managers/events";
 import { MapDisplayManager } from "@/engine/core/managers/map/MapDisplayManager";
 import {
@@ -48,7 +46,7 @@ import {
   TConditionList,
 } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { areObjectsOnSameLevel } from "@/engine/core/utils/object";
+import { areObjectsOnSameLevel } from "@/engine/core/utils/position";
 import { areCommunitiesEnemies, ERelation, setObjectSympathy } from "@/engine/core/utils/relation";
 import { canSquadHelpActor, updateSquadInvulnerabilityState } from "@/engine/core/utils/squad";
 import { vectorToString } from "@/engine/core/utils/vector";
@@ -60,7 +58,6 @@ import { SMART_TERRAIN_SECTION } from "@/engine/lib/constants/sections";
 import { FALSE, NIL, TRUE } from "@/engine/lib/constants/words";
 import {
   ALifeSmartTerrainTask,
-  AnyObject,
   ClientObject,
   LuaArray,
   NetPacket,
