+++ src/engine/core/schemes/restrictor/sr_teleport/TeleportManager.ts
@@ -7,7 +7,7 @@ import {
   ISchemeTeleportState,
 } from "@/engine/core/schemes/restrictor/sr_teleport/ISchemeTeleportState";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { teleportActorWithEffects } from "@/engine/core/utils/object";
+import { teleportActorWithEffects } from "@/engine/core/utils/position";
 import { trySwitchToAnotherSection } from "@/engine/core/utils/scheme/scheme_switch";
 import { postProcessors } from "@/engine/lib/constants/animation";
 import { ClientObject, Optional, TProbability, TTimestamp, Vector } from "@/engine/lib/types";
