+++ src/engine/core/schemes/stalker/help_wounded/actions/ActionHelpWounded.ts
@@ -6,7 +6,7 @@ import { EStalkerState } from "@/engine/core/objects/animation/types";
 import { ISchemeHelpWoundedState } from "@/engine/core/schemes/stalker/help_wounded/ISchemeHelpWoundedState";
 import { freeSelectedWoundedStalkerSpot } from "@/engine/core/schemes/stalker/help_wounded/utils";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { sendToNearestAccessibleVertex } from "@/engine/core/utils/object";
+import { sendToNearestAccessibleVertex } from "@/engine/core/utils/position";
 import { EClientObjectPath, Optional, TNumberId, Vector } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
