+++ src/engine/core/ui/debug/sections/DebugTreasuresSection.ts
@@ -6,7 +6,7 @@ import { ITreasureDescriptor, TreasureManager } from "@/engine/core/managers/tre
 import { AbstractDebugSection } from "@/engine/core/ui/debug/sections/AbstractDebugSection";
 import { isGameStarted } from "@/engine/core/utils/game";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isGameVertexFromLevel } from "@/engine/core/utils/object";
+import { isGameVertexFromLevel } from "@/engine/core/utils/position";
 import { EElementType, registerStatics, registerUiElement, resolveXmlFile } from "@/engine/core/utils/ui";
 import { Optional, ServerObject, TCount, TLabel, TNumberId, TPath, TSection, XmlInit } from "@/engine/lib/types";
