+++ src/engine/core/utils/scheme/scheme_switch.ts
@@ -5,7 +5,7 @@ import { abort, assert } from "@/engine/core/utils/assertion";
 import { pickSectionFromCondList } from "@/engine/core/utils/ini/ini_config";
 import { LuaLogger } from "@/engine/core/utils/logging";
 import { isActorSeenByObject } from "@/engine/core/utils/object/object_check";
-import { getDistanceBetween, isObjectInZone } from "@/engine/core/utils/object/object_location";
+import { getDistanceBetween, isObjectInZone } from "@/engine/core/utils/position";
 import { emitSchemeEvent } from "@/engine/core/utils/scheme/scheme_event";
 import { activateSchemeBySection } from "@/engine/core/utils/scheme/scheme_logic";
 import { NIL } from "@/engine/lib/constants/words";
