+++ src/engine/core/utils/squad/squad_actions.ts
@@ -1,7 +1,7 @@
 import { registry } from "@/engine/core/database";
 import { Squad } from "@/engine/core/objects/server/squad";
 import { hasAlifeInfo } from "@/engine/core/utils/info_portion";
-import { areObjectsOnSameLevel } from "@/engine/core/utils/object/object_location";
+import { areObjectsOnSameLevel } from "@/engine/core/utils/position";
 import { isEmpty } from "@/engine/core/utils/table";
 import { communities } from "@/engine/lib/constants/communities";
 import { infoPortions } from "@/engine/lib/constants/info_portions";
