+++ src/engine/scripts/declarations/effects/actor.ts
@@ -16,7 +16,8 @@ import type { Squad } from "@/engine/core/objects/server/squad";
 import { abort, assert, assertDefined } from "@/engine/core/utils/assertion";
 import { extern } from "@/engine/core/utils/binding";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isObjectInZone, objectPunchActor } from "@/engine/core/utils/object";
+import { objectPunchActor } from "@/engine/core/utils/object/object_action";
+import { isObjectInZone } from "@/engine/core/utils/position";
 import { giveItemsToActor } from "@/engine/core/utils/reward";
 import { detectorsOrder } from "@/engine/lib/constants/items/detectors";
 import { helmets } from "@/engine/lib/constants/items/helmets";
