+++ src/engine/scripts/declarations/effects/position.ts
@@ -5,7 +5,7 @@ import { Squad } from "@/engine/core/objects/server/squad";
 import { abort, assertDefined } from "@/engine/core/utils/assertion";
 import { extern } from "@/engine/core/utils/binding";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { isObjectInZone } from "@/engine/core/utils/object";
+import { isObjectInZone } from "@/engine/core/utils/position";
 import {
   ClientObject,
   Optional,
