+++ src/engine/scripts/declarations/effects/quests.ts
@@ -14,7 +14,7 @@ import { abort } from "@/engine/core/utils/assertion";
 import { extern, getExtern } from "@/engine/core/utils/binding";
 import { createGameAutoSave } from "@/engine/core/utils/game/game_save";
 import { disableInfo, giveInfo, hasAlifeInfo } from "@/engine/core/utils/info_portion";
-import { isObjectInZone } from "@/engine/core/utils/object";
+import { isObjectInZone } from "@/engine/core/utils/position";
 import { giveItemsToActor, takeItemFromActor } from "@/engine/core/utils/reward";
 import { spawnObject, spawnObjectInObject, spawnSquadInSmart } from "@/engine/core/utils/spawn";
 import { copyVector, createEmptyVector, createVector } from "@/engine/core/utils/vector";
